Bio_io

The full API is browsable here.

Bio_io is an OCaml library that provides programmatic access to common file formats used in bioinformatics like FASTA files.

If you have any issues or find any bugs, open an issue on the GitHub page.

License

Licensed under the Apache License, Version 2.0 or the MIT license, at your option. This program may not be copied, modified, or distributed except according to those terms.

Modules

These are the main modules of Bio_io. API docs and examples can be found on the respective pages.

Fasta_in_channel: an input channel for reading FASTA files

Fasta_record: a record type for FASTA files