Little_logger

A tiny, little logger <3

The full API is browsable here.

If you have any issues or find any bugs, open an issue on the GitHub page.

License

Licensed under the Apache License, Version 2.0 or the MIT license, at your option. This program may not be copied, modified, or distributed except according to those terms.

Modules

These are the main modules of Little_logger. API docs and examples can be found on the respective pages.

Logger: the main logger module